refactor: move indicator matrix fixtures to own conftest
